Avoid overflow

+++ crates/model/src/graph.rs
@@ -3,20 +3,20 @@ use specta::Type;
 
 #[derive(Deserialize, Type)]
 pub struct Answer {
-    pub path: Vec<u8>,
+    pub path: Vec<u32>,
     pub time_ms: u32,
 }
 
 #[derive(Clone, Deserialize, serde::Serialize, Type)]
 pub struct Edge {
-    pub source: u8,
-    pub target: u8,
+    pub source: u32,
+    pub target: u32,
 }
 
 #[derive(serde::Serialize, Type)]
 pub struct Graph {
     pub id: i64,
-    pub num_nodes: u8,
+    pub num_nodes: u32,
     pub edges: Vec<Edge>,
     pub best_time_ms: Option<u32>,
     pub cycle_found: bool,
